Added efficiency improvements
This version is atucally simpler. We need an uniform distribution from
0 to A, where A=N**L, N is the number of characters in the selection set,
L is the length of the password (we just manupulate it as
if we were dealing with a number in the base of N having L "digits").

We are trying to "fit" as many whole numbers A in the binary representation of
the random integer we read, discarding the "tail". To make the tail as small as
possible we add one extra byte if the ratio of the read number to A is smaller
than 128.

The previous version had issues with uniformity when I eliminated extra bits and also
with efficiency, as extra byte had to be often read, sometimes multiple times.